+++ app/commands/index.js
@@ -1,26 +1,26 @@
-import {
-  execChatAutoReplyOffCommand,
-  execChatAutoReplyOnCommand,
-  execChatCommand,
-  isChatAutoReplyOffCommand,
-  isChatAutoReplyOnCommand,
-  isChatCommand,
-} from './chat.js';
+import { execActivateCommand, isActivateCommand } from './activate.js';
+import { execChatCommand, isChatCommand } from './chat.js';
+import { isContinue } from './continue.js';
+import { execDeactivateCommand, isDeactivateCommand } from './deactivate.js';
 import { execDeployCommand, isDeployCommand } from './deploy.js';
 import { execDrawCommand, isDrawCommand } from './draw.js';
+import { execSettingsCommand, isSettings } from './settings.js';
 import { execVersionCommand, isVersionCommand } from './version.js';
 
 export {
-  execChatAutoReplyOffCommand,
-  execChatAutoReplyOnCommand,
+  execActivateCommand,
   execChatCommand,
+  execDeactivateCommand,
   execDeployCommand,
   execDrawCommand,
+  execSettingsCommand,
   execVersionCommand,
-  isChatAutoReplyOffCommand,
-  isChatAutoReplyOnCommand,
+  isActivateCommand,
   isChatCommand,
+  isContinue,
+  isDeactivateCommand,
   isDeployCommand,
   isDrawCommand,
+  isSettings,
   isVersionCommand,
 };