+++ tests/activate.test.js
@@ -4,7 +4,7 @@ import {
 import {
   getSession, handleEvents, removeSession, settings,
 } from '../app/index.js';
-import { COMMAND_CHAT, COMMAND_DEACTIVATE } from '../constants/command.js';
+import { COMMAND_ACTIVATE, COMMAND_DEACTIVATE } from '../constants/command.js';
 import storage from '../storage/index.js';
 import { createMessageEvents, TIMEOUT, USER_ID } from './utils.js';
 
@@ -16,10 +16,12 @@ afterEach(() => {
   removeSession(USER_ID);
 });
 
-test('COMMAND_CHAT', async () => {
+test('COMMAND_ACTIVATE', async () => {
   const events = [
     ...createMessageEvents([COMMAND_DEACTIVATE.text]),
-    ...createMessageEvents([`${COMMAND_CHAT.text} 嗨`]),
+    ...createMessageEvents(['嗨']), // should be ignored
+    ...createMessageEvents([COMMAND_ACTIVATE.text]),
+    ...createMessageEvents(['嗨']),
   ];
   let results;
   try {
@@ -32,6 +34,7 @@ test('COMMAND_CHAT', async () => {
   expect(replies).toEqual(
     [
       [COMMAND_DEACTIVATE.reply],
+      [COMMAND_ACTIVATE.reply],
       ['OK'],
     ],
   );