+++ tests/deactivate.test.js
@@ -4,7 +4,7 @@ import {
 import {
   getSession, handleEvents, removeSession, settings,
 } from '../app/index.js';
-import { COMMAND_CHAT, COMMAND_DEACTIVATE } from '../constants/command.js';
+import { COMMAND_DEACTIVATE } from '../constants/command.js';
 import storage from '../storage/index.js';
 import { createMessageEvents, TIMEOUT, USER_ID } from './utils.js';
 
@@ -16,10 +16,10 @@ afterEach(() => {
   removeSession(USER_ID);
 });
 
-test('COMMAND_CHAT', async () => {
+test('COMMAND_DEACTIVATE', async () => {
   const events = [
     ...createMessageEvents([COMMAND_DEACTIVATE.text]),
-    ...createMessageEvents([`${COMMAND_CHAT.text} 嗨`]),
+    ...createMessageEvents(['嗨']), // should be ignored
   ];
   let results;
   try {
@@ -27,12 +27,11 @@ test('COMMAND_CHAT', async () => {
   } catch (err) {
     console.error(err);
   }
-  expect(getSession(USER_ID).lines.length).toEqual(3 * 2);
+  expect(getSession(USER_ID).lines.length).toEqual(1 * 2);
   const replies = results.map(({ messages }) => messages.map(({ text }) => text));
   expect(replies).toEqual(
     [
       [COMMAND_DEACTIVATE.reply],
-      ['OK'],
     ],
   );
 }, TIMEOUT);