+++ tests/default.test.js
@@ -2,9 +2,9 @@ import {
   afterEach, beforeEach, expect, test,
 } from '@jest/globals';
 import {
-  getSession, handleEvents, removeSession, settings,
+  getSession, handleEvents, printSessions, removeSession, settings,
 } from '../app/index.js';
-import { COMMAND_CHAT, COMMAND_DEACTIVATE } from '../constants/command.js';
+import config from '../config/index.js';
 import storage from '../storage/index.js';
 import { createMessageEvents, TIMEOUT, USER_ID } from './utils.js';
 
@@ -16,10 +16,9 @@ afterEach(() => {
   removeSession(USER_ID);
 });
 
-test('COMMAND_CHAT', async () => {
+test('DEFAULT', async () => {
   const events = [
-    ...createMessageEvents([COMMAND_DEACTIVATE.text]),
-    ...createMessageEvents([`${COMMAND_CHAT.text} 嗨`]),
+    ...createMessageEvents(['嗨']),
   ];
   let results;
   try {
@@ -31,8 +30,8 @@ test('COMMAND_CHAT', async () => {
   const replies = results.map(({ messages }) => messages.map(({ text }) => text));
   expect(replies).toEqual(
     [
-      [COMMAND_DEACTIVATE.reply],
       ['OK'],
     ],
   );
+  if (config.APP_DEBUG) printSessions();
 }, TIMEOUT);