+++ utils/index.js
@@ -1,5 +1,3 @@
-import capitalize from './capitalize.js';
-import createAction from './create-action.js';
 import fetchEnvironment from './fetch-environment.js';
 import generateCompletion from './generate-completion.js';
 import generateImage from './generate-image.js';
@@ -8,8 +6,6 @@ import replyMessage from './reply-message.js';
 import validateSignature from './validate-signature.js';
 
 export {
-  capitalize,
-  createAction,
   fetchEnvironment,
   generateCompletion,
   generateImage,